+++ tests/testthat/helper.R
@@ -1,3 +1,10 @@
+skip_without_cargo <- function() {
+  skip_on_cran()
+  if (Sys.which("cargo") == "") {
+    skip("cargo not found on PATH")
+  }
+}
+
 skip_without_pyarrow <- function() {
   skip_on_cran()
   if (tolower(Sys.getenv("_R_CHECK_FORCE_SUGGESTS_")) != "false") {